+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Practic"
         tools:targetApi="31">
         <activity
-            android:name=".theme_1_kotlin.ru_git.task_1_shared.SharedPreferencesActivity"
+            android:name=".themes.theme_1_kotlin.ru_git.task_1_shared.SharedPreferencesActivity"
             android:exported="true">
 
             <intent-filter>
