--- app/src/main/java/com/example/practic/interview_questions/3_Collections.kt
@@ -1,17 +1,0 @@
-package com.example.practic.interview_questions
-
-import java.util.LinkedHashSet
-import java.util.LinkedList
-import java.util.Queue
-import java.util.SequencedCollection
-import java.util.TreeMap
-import java.util.TreeSet
-
-fun main() {
-    /*val list = java.util.HashMap
-    val list = Map
-    val list = Queue
-    val list = sortedSetOf()*/
-
-
-}
