+++ app/src/main/java/com/example/practic/themes/theme_1_kotlin/interview_questions/3_Collections.kt
@@ -1,0 +1,17 @@
+package com.example.practic.themes.theme_1_kotlin.interview_questions
+
+import java.util.LinkedHashSet
+import java.util.LinkedList
+import java.util.Queue
+import java.util.SequencedCollection
+import java.util.TreeMap
+import java.util.TreeSet
+
+fun main() {
+    /*val list = java.util.HashMap
+    val list = Map
+    val list = Queue
+    val list = sortedSetOf()*/
+
+
+}
