+++ app/src/main/java/com/example/practic/themes/theme_1_kotlin/interview_questions/Constructor.kt
