+++ app/src/main/java/com/example/practic/themes/theme_1_kotlin/interview_questions/Data_class.kt
@@ -1,0 +1,16 @@
+package com.example.practic.themes.theme_1_kotlin.interview_questions
+
+fun main() {
+}
+
+/**
+data class в Kotlin автоматически является final.
+Это означает, что от него нельзя наследоваться.
+
+но сам наследовать может */
+data class Data(val name: Int = 6) : Data2() {
+    val name3 = name.hashCode()
+}
+
+
+open class Data2(val name2: Int = 6)
