+++ app/src/main/java/com/example/practic/themes/theme_1_kotlin/practice_block/task_2.kt
