+++ app/src/main/java/com/example/practic/themes/theme_2_coroutines/interview_questions/CancellationException.kt
@@ -1,0 +1,19 @@
+package com.example.practic.themes.theme_2_coroutines.interview_questions
+
+/*
+fun com.example.practic.theme_2_coroutines.main() = runBlocking {
+    val job = launch {
+        try {
+            println("Корутину начинают работать")
+            delay(5000)  // Эмуляция долгой работы
+            println("Корутину завершили")
+        } catch (e: CancellationException) {
+            println("Корутину отменили: ${e.message}")
+        }
+    }
+
+    delay(2000)  // Ждём 2 секунды
+    job.cancel()  // Отменяем корутину
+    job.join()    // Дожидаемся завершения корутины
+}
+*/
