+++ app/src/main/java/com/example/practic/themes/theme_2_coroutines/interview_questions/Flow_try_catch.kt
@@ -1,0 +1,21 @@
+package com.example.practic.themes.theme_2_coroutines.interview_questions
+
+import kotlinx.coroutines.flow.*
+import kotlinx.coroutines.runBlocking
+
+fun main() = runBlocking {
+    flow {
+        emit(1)
+        emit(2)
+        emit(3)
+        throw Exception("An error occurred!")
+        emit(4)
+    }.collect { value ->
+        try {
+            if (value == 3) throw Exception("Error at value 3")
+            println("Received: $value")
+        } catch (e: Exception) {
+            println("Caught error: ${e.message}")
+        }
+    }
+}
