+++ app/src/main/java/com/example/practic/themes/theme_2_coroutines/interview_questions/yield.kt
@@ -1,0 +1,34 @@
+package com.example.practic.themes.theme_2_coroutines.interview_questions
+
+/*
+fun com.example.practic.theme_2_coroutines.main(): Unit = runBlocking {
+    launch {
+        repeat(5) {
+            println("A $it")
+            yield() // приостановить корутину, чтобы другая могла работать
+        }
+    }
+    launch {
+        repeat(5) {
+            println("B $it")
+            yield() // тоже приостановить свою корутину
+        }
+    }
+    launch {
+        repeat(5) {
+            println("C $it")
+            yield() // тоже приостановить свою корутину
+        }
+    }
+    launch {
+        repeat(5) {
+            println("D $it")
+            yield() // тоже приостановить свою корутину
+        }
+    }
+}
+
+class Frag2() : Fragment() {
+
+}
+*/
