+++ app/src/main/java/com/example/practic/themes/theme_2_coroutines/practice_block/LifecycleScope.kt
@@ -1,0 +1,19 @@
+package com.example.practic.themes.theme_2_coroutines.practice_block
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.lifecycleScope
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        lifecycleScope.launch {
+            delay(1000)
+            println("Task completed")
+        }
+    }
+}
